requirements.txt
streamlit
langchain
langchain-community
langchain-huggingface
langchain-text-splitters
sentence-transformers>=3.0.0
chromadb
pypdf
huggingface-hub
numpy>=1.26.0
scipy>=1.12.0
scikit-learn>=1.4.0
python-dotenv
